Refactor registering-proxy.ts to separate production code from test utilities
- Move MockRegistrar to src/lib-test/MockRegistrar.ts with proper vitest mocks
- Move Client interface to src/lib-test/fixtures.ts
- Update tests to use vitest spy functions instead of manual tracking
- Remove test utilities from production code
- All tests continue to pass

Resolves #2
